Temporary commit for rebase

+++ project/core/src/com/tdt4240/RawHeroes/view/uiElements/MainMenuButtonsFactory.java
@@ -24,7 +24,7 @@ public class MainMenuButtonsFactory{
         return createButton(text, xPos, yPos, GameConstants.BUTTON_WIDTH * 2, GameConstants.BUTTON_HEIGHT);
     }
 
-    public static TextButton createButton(String text, int xPos, int yPos, int xSize, int ySize) {
+    public static TextButton createButton(String text, float xPos, float yPos, float xSize, float ySize) {
 
         TextButton button = new TextButton(text, skin);
         button.setPosition(xPos, yPos);
